+++ src/icons/Main/CreateDeleteIcon.js
@@ -1,0 +1,18 @@
+export const deleteItemArrow = () => (
+    <svg width="24" height="26" viewBox="0 0 24 26" fill="none" xmlns="http://www.w3.org/2000/svg">
+<g opacity="0.75" filter="url(#filter0_d)">
+<path d="M19.5 8.13398C20.1667 8.51888 20.1667 9.48113 19.5 9.86603L6 17.6603C5.33334 18.0452 4.5 17.564 4.5 16.7942L4.5 1.20577C4.5 0.435972 5.33333 -0.0451548 6 0.339746L19.5 8.13398Z" fill="white"/>
+</g>
+<defs>
+<filter id="filter0_d" x="0.5" y="0.203125" width="23.5" height="25.5914" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
+<feFlood flood-opacity="0" result="BackgroundImageFix"/>
+<feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"/>
+<feOffset dy="4"/>
+<feGaussianBlur stdDeviation="2"/>
+<feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.45 0"/>
+<feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow"/>
+<feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow" result="shape"/>
+</filter>
+</defs>
+</svg>
+)
