+++ src/view/components/Create/Delete/DeleteItem.js
@@ -1,11 +1,12 @@
 import React from 'react'
 import styled from 'styled-components'
+import WorkspaceWaitItem from '../Workspace/WorkspaceWaitItem'
 
-const DeleteItem = () => {
+const DeleteItem = ({data}) => {
     
     return(
         <Container>
-            
+            <WorkspaceWaitItem data={data} />
         </Container>
     )
 }
