+++ src/view/components/Create/Workspace/WorkspaceMainItem.js
@@ -1,125 +1,104 @@
-import React, { useState } from 'react'
-import cn from 'classnames';
+import React from 'react'
 import styled from 'styled-components'
-import { locationIcon, openTimeIcon } from '../../../../icons/Main/CreateWorkspaceIcon';
+
 
 const WorkspaceMainItem = ({data}) => {
-
-    const [filter, setFliter] = useState(false)
-
     return(
-        <Container onMouseEnter={()=>{setFliter(true)}} onMouseLeave={()=>{setFliter(false)}}>
-            <WorkSpaceItem>
-              <ImageBox>
-                  <img src={data.background} alt="" />
-                  <div className={cn("filter",{isActive:filter})} />
-              </ImageBox>  
-              <TextArea className={cn({isActive:filter})}>
-                  <h1 className="title">{data.title}</h1>
-                  <span className="type">{data.type}</span>
-                  <p className="desc">{data.desc}</p>
-                  <div className="locationAndInfo">
-                    <span className="info">{openTimeIcon()}{data.info}</span>
-                    <span className="location">{locationIcon()}{data.location}</span>
-                  </div>
-              </TextArea>
-            </WorkSpaceItem>   
-            <ApprovedAt>
-                <span>
-                    {data.approvedAt}
-                </span>
-            </ApprovedAt>
+        <Container>
+            <WorkspaceItem>
+                <ImageBox>
+                    <img src={data.background} alt=""/>
+                </ImageBox>
+                
+                <ItemArea>
+                    <Text>
+                        <h1>
+                            {data.title}
+                        </h1>
+                        <span className={"type"}>
+                            {data.date}
+                        </span>
+                    </Text>
+                    <SubmitButton>
+                     <a href="#!">승인 요청</a>
+                    </SubmitButton>
+                </ItemArea>
+            </WorkspaceItem>
         </Container>
     )
 }
 
 const Container = styled.div`
-`
-const WorkSpaceItem = styled.div`
-    cursor: pointer;
-    padding: 240px 15px 0 15px;
-    margin-right: 20px;
-    width: 283px;
-    height: 460px;
-    box-sizing: border-box;
-    border: solid 1px rgba(255,255,255,0.15);
     position: relative;
+    margin-right: 20px;
+`
+const WorkspaceItem = styled.div`  
+    box-sizing:border-box;
+    width: 348px;
+    height: 224px;
+    border-radius: 20px;
+    background: #1F1F20;
     display: flex;
     flex-direction: column;
-    margin-bottom: 8px;
+    overflow: hidden;
 `
 const ImageBox = styled.div`
-    position: absolute;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    top: 0;
-    .filter{
-        position: absolute;
-        right: 0;
-        left: 0;
-        top: 0;
-        bottom: 0;
-        background: #37383A;
-        opacity: 0.45;
-        transition: all 0.3s;
-        &.isActive{
-            background: rgba(255, 255, 255, 0.479);
-        }
-    }
+    width: 100%;
+    height: 162px;
+    overflow: hidden;
+    cursor: pointer;
     img{
-        height: 100%;
         width: 100%;
+        height: 100%;
         object-fit: cover;
     }
 `
-const TextArea = styled.div`
-    color:#fff;
-    position: relative;
-    h1{
-        
-        opacity: 0.9;
-        font-size: 21px;
-        font-weight: bold;
-        margin-bottom: 10px;
-    }
-    svg{
-        margin-right: 5px;
-    }
-    &.isActive{
+const ItemArea = styled.div`
+    padding: 0 15px;
+    width: 100%;
+    box-sizing: border-box; 
+    display: flex;
+    align-items: center;
+    flex: 1;
+    justify-content: space-between;
+ 
+`
+const Text = styled.div`
+   h1{
+        font-size: 15px;
+        line-height: 18px;
+        opacity: 0.8;
+        margin-bottom: 2px;
+        cursor: pointer;
     }
     .type{
-        
+        color:#fff;
         opacity: 0.45;
         font-size: 9px;
-        margin-bottom: 20px;
+        display: block;
     }
     .desc{
-        margin-top: 15px;
-        font-size: 9px;
+        font-size: 11px;
+        color:#fff;
         opacity: 0.8;
-        margin-bottom: 15px;
-        line-height: 1.8;
-    }
-    .info{
-        font-size: 9px;
-        
-        opacity: 0.8;
-        margin-bottom: 7px;
-    }
-    .location{
-        font-size: 9px;
-        opacity: 0.8;
-    }
-    .locationAndInfo{
-        display: flex;
-        flex-direction: column;
     }
 `
-const ApprovedAt = styled.div`
-    span{
-        color:#71777C;
-        font-size: 9px;
+
+const SubmitButton = styled.div`
+    position: relative;
+    display: flex;
+    align-items: center;
+    width: 69px;
+    justify-content: center;
+    height:28px;
+    border-radius: 15px;
+    border:solid 1px #4F80FF;
+    a{
+        color:#4F80FF;
+        font-size: 11px;
+        text-decoration: none;
+        display: block;
+        
     }
 `
